requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-cors-headers==3.4.0
django-filter==2.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
simplejson==3.17.0
sqlparse==0.3.1
urllib3==1.25.11
zipp==3.1.0
